feat: createDateAndTime module created
