+++ routes/messages/messages.js
@@ -1,5 +1,6 @@
 const express = require('express');
 const router =  express.Router();
+const createDateAndTime = require('../../utils/createDateAndTime');
 
 const welcomeMessage = {
   id: 0,
@@ -79,13 +80,5 @@ router.put('/update/:id', function (req, res) {
         });
     }
 })
-
-function createDateAndTime () {
-    const today = new Date();
-    const date = `${today.getFullYear()}-${today.getMonth() + 1}-${today.getDate()}`;
-    const time = `${today.getHours()}:${today.getMinutes()}:${today.getSeconds()}`;
-    const dateTime = `${date} ${time}`;
-    return dateTime;
-}
 
 module.exports = router;
