+++ utils/createDateAndTime.js
@@ -1,0 +1,10 @@
+
+function createDateAndTime () {
+    const today = new Date();
+    const date = `${today.getFullYear()}-${today.getMonth() + 1}-${today.getDate()}`;
+    const time = `${today.getHours()}:${today.getMinutes()}:${today.getSeconds()}`;
+    const dateTime = `${date} ${time}`;
+    return dateTime;
+}
+
+module.exports = createDateAndTime;
